[RENAME] VC명 변경
