--- Instagram/Instagram/SuccessViewController.swift
@@ -1,28 +1,0 @@
-//
-//  SuccessViewController.swift
-//  Instagram
-//
-//  Created by 변주현 on 2022/04/07.
-//
-
-import UIKit
-
-class SuccessViewController: UIViewController {
-
-    @IBOutlet weak var welcomeLabel: UILabel!
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-    }
-    
-    @IBAction func okButton(_ sender: Any) {
-        let presentingVC = self.presentingViewController!
-        let navigationController = presentingVC is UINavigationController ? presentingVC as? UINavigationController : presentingVC.navigationController
-        
-        self.dismiss(animated: true){
-            navigationController?.popToRootViewController(animated: true)
-            return
-        }
-    }
-    
-}
